API update

+++ core/src/main/java/org/mwdb/chunk/KChunkIterator.java
@@ -6,6 +6,6 @@ public interface KChunkIterator {
 
     KChunk next();
 
-    int size();
+    long size();
 
 }
