+++ core/src/main/java/org/mwdb/chunk/KStack.java
@@ -2,12 +2,10 @@ package org.mwdb.chunk;
 
 public interface KStack {
 
-    boolean enqueue(int index);
+    boolean enqueue(long index);
 
     long dequeueTail();
 
-    boolean dequeue(int index);
-
-    // void reenqueue(int index);
+    boolean dequeue(long index);
 
 }
